LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/adding-snap-area-while-snapped-expected.txt


FAIL Adding a new snap area while already snapped should not make the scroller snap to it. assert_equals: expected 100 but got 0
